Remove some code that was supposed to be dead
Some code used to be in an
if (psrtpriv->silent_reset_inprogress == _TRUE)
conditional, which was always false as silent_reset_inprogress
didn't get set by anything. So this was never called before we
disabled and removed the DBG_CONFIG_ERROR_DETECT option in commit
87db1e75.

With help from Jaedon Shin.

Closes #5