rtl8723bs: Fix build broblem
In lernel 4.2, the calling sequence for cfg80211_connect_result() will
change. Unfortunately, at the moment the new code exists only in the
wireless-drivers-next repo, but the kernel version is still 4.1. As
I use that repo for my regular kernel, I have to make a local
modification. The problem comes because I committed and pushed the
wrong version. I will try to keep it straight.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>