rtl8723bs: Remove _func_enter_ and _func_exit_ macros from core/*.c
As the kernel does a good job of profilinf functions, these macros can be
removed. In a few instances, removal of the exit macro left a label immediately
preceeding the closing brace of the function. With gcc, this is an error, thus
for these routines, any goto's referencing that label were converted to return
and the label was deleted.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>