rtl8723bs: Fix kernel patches
In kernel 4.4, file drivers/mmc/host/sdhci-pci.c was renamed to
drivers/mmc/host/sdhci-pci-core.c. This change breaks patch #4. To fix
that, #4 was renamed to patches/0004a-mmc-sdhci-pci-Fix-device-hang-on-Intel-BayTrail.patch
and a new patch patches/0004b-mmc-sdhci-pci-Fix-device-hang-on-Intel-BayTrail.patch was
created that accounts for the name change.

Patch #4a is to be used for kernels < 4.4.0. Use #4b for 4.4.0 and later.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>