Fix antenna-path for rtl8703as
The rtl8703as in reality is a rtl8723bs in a different package.
(it has the same sdio-id).

In android the rtl8703as uses the rtl8723bs-vq0 driver instead of the
regular rtl8732bs driver. Looking at a diff between the 2 the only
difference is that the vq0 driver forces the antenna-path to
ODM_RF_PATH_B.

It seems that we can identify the rtl8703as by its package, as it comes
in a QNF68 package, where as the rtl8723bs comes in a TFBGA79 package,
so we can simply support both in the same driver.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>