Kill CONFIG_RTW_SDIO_PM_KEEP_POWER
As the workaround mentions this is for android-4.0 only, so we don't
really need this. Moreover it seems to only kick in after the first
suspend/resume cycle (likely the flag should be set earlier the in
the suspend callback to make it work in the first cycle).

On some of my hardware this workaround results in wifi not being
available after the first suspend/resume, so lets remove it.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>