Fix the module getting loaded on systems with an unsupported sdio-wifi device
Using SDIO_DEVICE_CLASS(SDIO_CLASS_WLAN) in our MODULE_DEVICE_TABLE
is of little use if we then check for specific vendor and product
functions in probe().

Actually list the supported ids in the MODULE_DEVICE_TABLE and remove
the now unnecessary check from probe().

This fixes udev unnecessarily loading the module on systems with
a sdio-wifi device not supported by this driver.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>