Migrate to staging driver
The driver is available upstream now in the linux-next tree, in
the staging drivers section. All tests, bug reports should be
made to the linux-wireless mailing-list, using this upstream driver.

Consider the existence of this GitHub repository as historical
documentation. Patches that were shipped as part of this repository have
mostly been subsided and obsoleted by upstream kernel fixes.

Support for Bluetooth, and the firmware files are also in the process
of being made available upstream, and will be removed when that happens.